+++ R/tadaa_plots.R
@@ -21,54 +21,68 @@
 #' @family Tadaa-plot functions
 #' @import ggplot2
 #' @examples
-#' \dontrun{
 #' tadaa_int(ngo, stunzahl, jahrgang, geschl)
 #'
 #' # As grid, if cowplot is installed
 #' tadaa_int(ngo, stunzahl, jahrgang, geschl, grid = TRUE)
-#' }
 tadaa_int <- function(data, response, group1, group2, grid = FALSE,
                       brewer_palette = "Set1", labels = c("A", "B"),
                       show_n = FALSE, print = TRUE) {
-
   if (show_n) {
     subtitle <- paste0("N = ", nrow(data))
   } else {
     subtitle <- NULL
   }
 
-  sdots <- lazyeval::interp(~mean(variable, na.rm = T), variable = substitute(response))
+  title1 <- ifelse(!grid, paste0(
+    "Interaction of ",
+    substitute(group1), " & ", substitute(group2)
+  ),
+  paste0(
+    "Interaction of\n",
+    substitute(group1), " & ", substitute(group2)
+  )
+  )
 
-  data <- dplyr::group_by_(data, substitute(group1), substitute(group2))
-  data <- dplyr::summarize_(data, .dots = list(mw = sdots))
+  title2 <- ifelse(!grid, paste0(
+    "Interaction of ",
+    substitute(group2), " & ", substitute(group1)
+  ),
+  paste0(
+    "Interaction of\n",
+    substitute(group2), " & ", substitute(group1)
+  )
+  )
 
-  title1 <- ifelse(!grid, paste0("Interaction of ",
-                                 substitute(group1), " & ", substitute(group2)),
-                   paste0("Interaction of\n",
-                          substitute(group1), " & ", substitute(group2)))
+  p1 <- ggplot(data = data, aes_string(
+    x = substitute(group1), y = substitute(response),
+    colour = substitute(group2)
+  )) +
+    stat_summary(aes_string(group = substitute(group2)),
+                 fun.y = mean, geom = "line") +
+    stat_summary(aes_string(group = substitute(group2)),
+                 fun.y = mean, geom = "point", shape = 23, fill = "white") +
+    scale_colour_brewer(palette = brewer_palette) +
+    labs(
+      title = title1, y = paste0("Mean ", substitute(response)),
+      subtitle = subtitle
+    ) +
+    theme(legend.position = "top")
 
-  title2 <- ifelse(!grid, paste0("Interaction of ",
-                                 substitute(group2), " & ", substitute(group1)),
-                   paste0("Interaction of\n",
-                          substitute(group2), " & ", substitute(group1)))
-
-  p1 <- ggplot(data = data, aes_string(x = substitute(group1), y = "mw",
-                                       colour = substitute(group2))) +
-          geom_line(aes_string(group = substitute(group2))) +
-          geom_point(shape = 23, fill = "white") +
-          scale_colour_brewer(palette = brewer_palette) +
-          labs(title = title1, y = paste0("Mean ", substitute(response)),
-               subtitle = subtitle) +
-          theme(legend.position = "top")
-
-  p2 <- ggplot(data = data, aes_string(x = substitute(group2), y = "mw",
-                                       colour = substitute(group1))) +
-          geom_line(aes_string(group = substitute(group1))) +
-          geom_point(shape = 23, fill = "white") +
-          scale_colour_brewer(palette = brewer_palette) +
-          labs(title = title2, y = paste0("Mean ", substitute(response)),
-               subtitle = subtitle) +
-          theme(legend.position = "top")
+  p2 <- ggplot(data = data, aes_string(
+    x = substitute(group2), y = substitute(response),
+    colour = substitute(group1)
+  )) +
+    stat_summary(aes_string(group = substitute(group1)),
+                 fun.y = mean, geom = "line") +
+    stat_summary(aes_string(group = substitute(group1)),
+                 fun.y = mean, geom = "point", shape = 23, fill = "white") +
+    scale_colour_brewer(palette = brewer_palette) +
+    labs(
+      title = title2, y = paste0("Mean ", substitute(response)),
+      subtitle = subtitle
+    ) +
+    theme(legend.position = "top")
 
   if (print) {
     if (!grid) {
@@ -103,41 +117,43 @@ tadaa_int <- function(data, response, group1, group2, grid = FALSE,
 #' @examples
 #' tadaa_balance(ngo, jahrgang, geschl)
 tadaa_balance <- function(data, group1, group2, palette = "D", annotate = TRUE) {
-
   group1 <- deparse(substitute(group1))
   group2 <- deparse(substitute(group2))
-  # group3 <- deparse(substitute(group3))
 
   heat <- table(data[[group1]], data[[group2]])
   heat <- as.data.frame(heat)
 
   if (annotate) {
-    anno <- geom_label(aes_string(label = "Freq"), stat = "identity",
-                       fill = "white", alpha = .5, size = 5)
+    anno <- geom_label(
+      aes_string(label = "Freq"), stat = "identity",
+      fill = "white", alpha = .5, size = 5
+    )
   } else {
     anno <- NULL
   }
 
-  balance <- ggplot(heat, aes_string(x = "Var1", y = "Var2", fill = "Freq")) +
+  ggplot(heat, aes_string(x = "Var1", y = "Var2", fill = "Freq")) +
     geom_tile(color = "white", size = 0.75) +
     anno +
-    labs(title = paste("Design Balance for", substitute(group1),
-                       "and", substitute(group2)),
-         x = group1, y = group2) +
+    labs(
+      title = paste(
+        "Design Balance for", substitute(group1),
+        "and", substitute(group2)
+      ),
+      x = group1, y = group2
+    ) +
     scale_x_discrete() +
     viridis::scale_fill_viridis(option = palette) +
     coord_equal(ratio = .625) +
     theme(legend.position = "none")
-
-  balance
 }
 
-#' Means with Errorbars
+#' Plot Means with Errorbars
 #'
 #' @param data A `data.frame`
 #' @param response Response variable, numeric.
 #' @param group Grouping variable, ideally a `factor`.
-#' @param brewer_palette Optional: The name of the `RColorBrewer`` palette to use,
+#' @param brewer_palette Optional: The name of the `RColorBrewer` palette to use,
 #' defaults to `Set1`. Use `NULL` for no brewer palette.
 #' @return A ggplot2 object.
 #' @export
@@ -146,15 +162,14 @@ tadaa_balance <- function(data, group1, group2, palette = "D", annotate = TRUE) 
 #' @examples
 #' tadaa_mean_ci(ngo, deutsch, jahrgang, brewer_palette = "Set1")
 tadaa_mean_ci <- function(data, response, group, brewer_palette = "Set1") {
-
   x <- deparse(substitute(group))
   y <- deparse(substitute(response))
 
   p <- ggplot(data = data, aes_string(x = x, y = y, color = x)) +
-        stat_summary(fun.data = "mean_ci_t", geom = "errorbar", width = 0.6, size = 1.5) +
-        stat_summary(fun.y = "mean", geom = "point", size = 3, color = "black") +
-        stat_summary(fun.y = "mean", geom = "point", size = 2, color = "white") +
-        guides(color = F)
+    stat_summary(fun.data = "mean_ci_t", geom = "errorbar", width = 0.6, size = 1.5) +
+    stat_summary(fun.y = "mean", geom = "point", size = 3, color = "black") +
+    stat_summary(fun.y = "mean", geom = "point", size = 2, color = "white") +
+    guides(color = F)
   if (!is.null(brewer_palette)) {
     p <- p + scale_color_brewer(palette = brewer_palette)
   }
@@ -173,7 +188,6 @@ tadaa_mean_ci <- function(data, response, group, brewer_palette = "Set1") {
 #' @return A [ggplot2] object.
 #' @export
 #' @import ggplot2
-#' @importFrom dplyr arrange
 #' @family Tadaa-plot functions
 #' @note The `alpha` of the error bars is set to `0.25` if the comparison
 #' is not significant, and `1` otherwise. That's neat.
@@ -182,32 +196,36 @@ tadaa_mean_ci <- function(data, response, group, brewer_palette = "Set1") {
 #' tadaa_plot_tukey(tests)
 tadaa_plot_tukey <- function(data, brewer_palette = "Set1") {
 
-  # Please R CMD check
-  comparison <- NULL; estimate <- NULL; conf.low <- NULL
-  conf.high <- NULL; term <- NULL; signif <- NULL
-
   data$signif <- ifelse(data$conf.high > 0 & data$conf.low < 0, "no", "yes")
 
-  data <- dplyr::arrange(data, term, estimate)
-  data$comparison <- factor(data$comparison,
-                            levels = rev(as.character(data$comparison)),
-                            ordered = TRUE)
+  data <- data[order(data$term, data$estimate), ]
+  data$comparison <- factor(
+    data$comparison,
+    levels = rev(as.character(data$comparison)),
+    ordered = TRUE
+  )
 
-  p <- ggplot(data = data,
-              aes(x     = comparison,
-                  y     = estimate,
-                  ymin  = conf.low,
-                  ymax  = conf.high,
-                  color = term,
-                  alpha = signif)) +
+  p <- ggplot(
+    data = data,
+    aes_string(
+      x = "comparison",
+      y = "estimate",
+      ymin = "conf.low",
+      ymax = "conf.high",
+      color = "term",
+      alpha = "signif"
+    )
+  ) +
     geom_errorbar(width = .75, size = 1.25) +
     geom_point(shape = 23, size = 1.5) +
     geom_hline(yintercept = 0, linetype = "dashed") +
     coord_flip() +
     scale_alpha_manual(values = c("no" = 0.25, "yes" = 1), guide = F) +
-    labs(title = "Tukey HSD Results", subtitle = "Mean Difference with 95% CI",
-         x = "Compared Groups", y = "Mean Difference", color = "Term (Factor)",
-         caption = "Confidence intervals not including x = 0 are considered significant") +
+    labs(
+      title = "Tukey HSD Results", subtitle = "Mean Difference with 95% CI",
+      x = "Compared Groups", y = "Mean Difference", color = "Term (Factor)",
+      caption = "Confidence intervals not including x = 0 are considered significant"
+    ) +
     theme(legend.position = "top")
 
   if (!is.null(brewer_palette)) {